requirements.txt
slack_bolt
python-dotenv
python-json-logger
fastapi
uvicorn
pytest
